backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
channels==4.1.0
daphne==4.1.2
Pillow==10.4.0
requests>=2.32
dnspython>=2.6
beautifulsoup4>=4.12
